third_party/WebKit/LayoutTests/inspector-protocol/css/css-getLayoutTreeNodes-expected.txt

Some Text
And More Text 
Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque sit amet est sem. Aenean ut neque volutpat, posuere odio at, mollis nibh. Aenean sodales nulla et ligula efficitur sollicitudin blandit sed lectus. Duis orci enim, sodales ac lectus sed, hendrerit efficitur est. Quisque gravida facilisis viverra.

Lets have a span
Rotated text!


LayoutTreeNodes result:
{
  "layoutTreeNodes": [
    {
      "nodeId": 1,
      "boundingBox": {
        "x": 0,
        "y": 0,
        "width": 785,
        "height": 600
      }
    },
    {
      "nodeId": 2,
      "boundingBox": {
        "x": 0,
        "y": 0,
        "width": 785,
        "height": 911
      },
      "styleIndex": 0
    },
    {
      "nodeId": 9,
      "boundingBox": {
        "x": 8,
        "y": 8,
        "width": 769,
        "height": 882
      },
      "styleIndex": 1
    },
    {
      "nodeId": 10,
      "boundingBox": {
        "x": 8,
        "y": 8,
        "width": 769,
        "height": 882
      },
      "styleIndex": 1
    },
    {
      "nodeId": 11,
      "boundingBox": {
        "x": 8,
        "y": 8,
        "width": 769,
        "height": 16
      },
      "styleIndex": 2
    },
    {
      "nodeId": 12,
      "boundingBox": {
        "x": 8,
        "y": 8,
        "width": 144,
        "height": 16
      },
      "layoutText": "Some Text",
      "inlineTextNodes": [
        {
          "boundingBox": {
            "x": 8,
            "y": 8,
            "width": 144,
            "height": 16
          },
          "startCharacterIndex": 0,
          "numCharacters": 9
        }
      ],
      "styleIndex": 3
    },
    {
      "nodeId": 13,
      "boundingBox": {
        "x": 8,
        "y": 568,
        "width": 224,
        "height": 16
      },
      "layoutText": " And More Text\n    ",
      "inlineTextNodes": [
        {
          "boundingBox": {
            "x": 8,
            "y": 568,
            "width": 224,
            "height": 16
          },
          "startCharacterIndex": 1,
          "numCharacters": 14
        }
      ],
      "styleIndex": 3
    },
    {
      "nodeId": 14,
      "boundingBox": {
        "x": 232,
        "y": 24,
        "width": 200,
        "height": 576
      },
      "styleIndex": 4
    },
    {
      "nodeId": 15,
      "boundingBox": {
        "x": 232,
        "y": 40,
        "width": 200,
        "height": 480
      },
      "styleIndex": 5
    },
    {
      "nodeId": 16,
      "boundingBox": {
        "x": 232,
        "y": 40,
        "width": 192,
        "height": 480
      },
      "layoutText": "\n        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque sit amet est sem.\n        Aenean ut neque volutpat, posuere odio at, mollis nibh. Aenean sodales nulla et\n        ligula efficitur sollicitudin blandit sed lectus. Duis orci enim, sodales ac lectus sed,\n        hendrerit efficitur est. Quisque gravida facilisis viverra.\n        ",
      "inlineTextNodes": [
        {
          "boundingBox": {
            "x": 232,
            "y": 40,
            "width": 176,
            "height": 16
          },
          "startCharacterIndex": 9,
          "numCharacters": 11
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 56,
            "width": 144,
            "height": 16
          },
          "startCharacterIndex": 21,
          "numCharacters": 9
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 72,
            "width": 80,
            "height": 16
          },
          "startCharacterIndex": 31,
          "numCharacters": 5
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 88,
            "width": 176,
            "height": 16
          },
          "startCharacterIndex": 37,
          "numCharacters": 11
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 104,
            "width": 160,
            "height": 16
          },
          "startCharacterIndex": 49,
          "numCharacters": 10
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 120,
            "width": 80,
            "height": 16
          },
          "startCharacterIndex": 60,
          "numCharacters": 5
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 136,
            "width": 192,
            "height": 16
          },
          "startCharacterIndex": 66,
          "numCharacters": 12
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 152,
            "width": 192,
            "height": 16
          },
          "startCharacterIndex": 79,
          "numCharacters": 12
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 168,
            "width": 80,
            "height": 16
          },
          "startCharacterIndex": 92,
          "numCharacters": 5
        },
        {
          "boundingBox": {
            "x": 312,
            "y": 168,
            "width": 96,
            "height": 16
          },
          "startCharacterIndex": 105,
          "numCharacters": 6
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 184,
            "width": 128,
            "height": 16
          },
          "startCharacterIndex": 112,
          "numCharacters": 8
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 200,
            "width": 144,
            "height": 16
          },
          "startCharacterIndex": 121,
          "numCharacters": 9
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 216,
            "width": 192,
            "height": 16
          },
          "startCharacterIndex": 131,
          "numCharacters": 12
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 232,
            "width": 160,
            "height": 16
          },
          "startCharacterIndex": 144,
          "numCharacters": 10
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 248,
            "width": 192,
            "height": 16
          },
          "startCharacterIndex": 155,
          "numCharacters": 12
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 264,
            "width": 112,
            "height": 16
          },
          "startCharacterIndex": 168,
          "numCharacters": 7
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 280,
            "width": 128,
            "height": 16
          },
          "startCharacterIndex": 176,
          "numCharacters": 8
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 296,
            "width": 96,
            "height": 16
          },
          "startCharacterIndex": 193,
          "numCharacters": 6
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 312,
            "width": 144,
            "height": 16
          },
          "startCharacterIndex": 200,
          "numCharacters": 9
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 328,
            "width": 192,
            "height": 16
          },
          "startCharacterIndex": 210,
          "numCharacters": 12
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 344,
            "width": 176,
            "height": 16
          },
          "startCharacterIndex": 223,
          "numCharacters": 11
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 360,
            "width": 192,
            "height": 16
          },
          "startCharacterIndex": 235,
          "numCharacters": 12
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 376,
            "width": 160,
            "height": 16
          },
          "startCharacterIndex": 248,
          "numCharacters": 10
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 392,
            "width": 160,
            "height": 16
          },
          "startCharacterIndex": 259,
          "numCharacters": 10
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 408,
            "width": 176,
            "height": 16
          },
          "startCharacterIndex": 270,
          "numCharacters": 11
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 424,
            "width": 144,
            "height": 16
          },
          "startCharacterIndex": 290,
          "numCharacters": 9
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 440,
            "width": 144,
            "height": 16
          },
          "startCharacterIndex": 300,
          "numCharacters": 9
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 456,
            "width": 192,
            "height": 16
          },
          "startCharacterIndex": 310,
          "numCharacters": 12
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 472,
            "width": 112,
            "height": 16
          },
          "startCharacterIndex": 323,
          "numCharacters": 7
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 488,
            "width": 144,
            "height": 16
          },
          "startCharacterIndex": 331,
          "numCharacters": 9
        },
        {
          "boundingBox": {
            "x": 232,
            "y": 504,
            "width": 128,
            "height": 16
          },
          "startCharacterIndex": 341,
          "numCharacters": 8
        }
      ],
      "styleIndex": 3
    },
    {
      "nodeId": 17,
      "boundingBox": {
        "x": 232,
        "y": 536,
        "width": 200,
        "height": 48
      },
      "styleIndex": 6
    },
    {
      "nodeId": 18,
      "boundingBox": {
        "x": 272,
        "y": 536,
        "width": 160,
        "height": 16
      },
      "styleIndex": 7
    },
    {
      "nodeId": 19,
      "boundingBox": {
        "x": 272,
        "y": 552,
        "width": 144,
        "height": 32
      },
      "styleIndex": 8
    },
    {
      "nodeId": 20,
      "boundingBox": {
        "x": 272,
        "y": 552,
        "width": 144,
        "height": 32
      },
      "layoutText": "Lets have a span",
      "inlineTextNodes": [
        {
          "boundingBox": {
            "x": 272,
            "y": 552,
            "width": 144,
            "height": 16
          },
          "startCharacterIndex": 0,
          "numCharacters": 9
        },
        {
          "boundingBox": {
            "x": 272,
            "y": 568,
            "width": 96,
            "height": 16
          },
          "startCharacterIndex": 10,
          "numCharacters": 6
        }
      ],
      "styleIndex": 8
    },
    {
      "nodeId": 21,
      "boundingBox": {
        "x": 92,
        "y": 516,
        "width": 32,
        "height": 200
      },
      "styleIndex": 9
    },
    {
      "nodeId": 22,
      "boundingBox": {
        "x": 92,
        "y": 516,
        "width": 32,
        "height": 112
      },
      "layoutText": "Rotated text!",
      "inlineTextNodes": [
        {
          "boundingBox": {
            "x": 108,
            "y": 516,
            "width": 16,
            "height": 112
          },
          "startCharacterIndex": 0,
          "numCharacters": 7
        },
        {
          "boundingBox": {
            "x": 92,
            "y": 516,
            "width": 16,
            "height": 80
          },
          "startCharacterIndex": 8,
          "numCharacters": 5
        }
      ],
      "styleIndex": 10
    },
    {
      "nodeId": 23,
      "boundingBox": {
        "x": 8,
        "y": 632,
        "width": 404,
        "height": 204
      },
      "styleIndex": 11
    },
    {
      "nodeId": 25,
      "boundingBox": {
        "x": 8,
        "y": 857,
        "width": 769,
        "height": 33
      },
      "styleIndex": 12
    }
  ],
  "computedStyles": [
    {
      "properties": [
        {
          "name": "transform",
          "value": "none"
        },
        {
          "name": "transform-origin",
          "value": "392.5px 455.438px"
        },
        {
          "name": "height",
          "value": "910.875px"
        },
        {
          "name": "width",
          "value": "785px"
        },
        {
          "name": "display",
          "value": "block"
        },
        {
          "name": "outline-color",
          "value": "rgb(0, 0, 0)"
        }
      ]
    },
    {
      "properties": [
        {
          "name": "transform",
          "value": "none"
        },
        {
          "name": "transform-origin",
          "value": "384.5px 440.719px"
        },
        {
          "name": "height",
          "value": "881.438px"
        },
        {
          "name": "width",
          "value": "769px"
        },
        {
          "name": "display",
          "value": "block"
        },
        {
          "name": "outline-color",
          "value": "rgb(0, 0, 0)"
        }
      ]
    },
    {
      "properties": [
        {
          "name": "transform",
          "value": "none"
        },
        {
          "name": "transform-origin",
          "value": "384.5px 8px"
        },
        {
          "name": "height",
          "value": "16px"
        },
        {
          "name": "width",
          "value": "769px"
        },
        {
          "name": "display",
          "value": "block"
        },
        {
          "name": "outline-color",
          "value": "rgb(0, 0, 0)"
        }
      ]
    },
    {
      "properties": [
        {
          "name": "transform",
          "value": "none"
        },
        {
          "name": "transform-origin",
          "value": "0px 0px"
        },
        {
          "name": "height",
          "value": "auto"
        },
        {
          "name": "width",
          "value": "auto"
        },
        {
          "name": "display",
          "value": "block"
        },
        {
          "name": "outline-color",
          "value": "rgb(0, 0, 0)"
        }
      ]
    },
    {
      "properties": [
        {
          "name": "transform",
          "value": "none"
        },
        {
          "name": "transform-origin",
          "value": "100px 288px"
        },
        {
          "name": "height",
          "value": "576px"
        },
        {
          "name": "width",
          "value": "200px"
        },
        {
          "name": "display",
          "value": "inline-block"
        },
        {
          "name": "outline-color",
          "value": "rgb(0, 0, 0)"
        }
      ]
    },
    {
      "properties": [
        {
          "name": "transform",
          "value": "none"
        },
        {
          "name": "transform-origin",
          "value": "100px 240px"
        },
        {
          "name": "height",
          "value": "480px"
        },
        {
          "name": "width",
          "value": "200px"
        },
        {
          "name": "display",
          "value": "block"
        },
        {
          "name": "outline-color",
          "value": "rgb(0, 0, 0)"
        }
      ]
    },
    {
      "properties": [
        {
          "name": "transform",
          "value": "none"
        },
        {
          "name": "transform-origin",
          "value": "100px 24px"
        },
        {
          "name": "height",
          "value": "48px"
        },
        {
          "name": "width",
          "value": "160px"
        },
        {
          "name": "display",
          "value": "block"
        },
        {
          "name": "outline-color",
          "value": "rgb(0, 0, 0)"
        }
      ]
    },
    {
      "properties": [
        {
          "name": "transform",
          "value": "none"
        },
        {
          "name": "transform-origin",
          "value": "80px 8px"
        },
        {
          "name": "height",
          "value": "16px"
        },
        {
          "name": "width",
          "value": "160px"
        },
        {
          "name": "display",
          "value": "list-item"
        },
        {
          "name": "outline-color",
          "value": "rgb(0, 0, 0)"
        }
      ]
    },
    {
      "properties": [
        {
          "name": "transform",
          "value": "none"
        },
        {
          "name": "transform-origin",
          "value": "0px 0px"
        },
        {
          "name": "height",
          "value": "auto"
        },
        {
          "name": "width",
          "value": "auto"
        },
        {
          "name": "display",
          "value": "inline"
        },
        {
          "name": "outline-color",
          "value": "rgb(0, 0, 0)"
        }
      ]
    },
    {
      "properties": [
        {
          "name": "transform",
          "value": "matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)"
        },
        {
          "name": "transform-origin",
          "value": "100px 16px"
        },
        {
          "name": "height",
          "value": "32px"
        },
        {
          "name": "width",
          "value": "200px"
        },
        {
          "name": "display",
          "value": "block"
        },
        {
          "name": "outline-color",
          "value": "rgb(0, 0, 0)"
        }
      ]
    },
    {
      "properties": [
        {
          "name": "transform",
          "value": "matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)"
        },
        {
          "name": "transform-origin",
          "value": "0px 0px"
        },
        {
          "name": "height",
          "value": "auto"
        },
        {
          "name": "width",
          "value": "auto"
        },
        {
          "name": "display",
          "value": "block"
        },
        {
          "name": "outline-color",
          "value": "rgb(0, 0, 0)"
        }
      ]
    },
    {
      "properties": [
        {
          "name": "transform",
          "value": "none"
        },
        {
          "name": "transform-origin",
          "value": "202px 102px"
        },
        {
          "name": "height",
          "value": "200px"
        },
        {
          "name": "width",
          "value": "400px"
        },
        {
          "name": "display",
          "value": "inline"
        },
        {
          "name": "outline-color",
          "value": "rgb(0, 0, 0)"
        }
      ]
    },
    {
      "properties": [
        {
          "name": "transform",
          "value": "none"
        },
        {
          "name": "transform-origin",
          "value": "384.5px 16px"
        },
        {
          "name": "height",
          "value": "32px"
        },
        {
          "name": "width",
          "value": "769px"
        },
        {
          "name": "display",
          "value": "block"
        },
        {
          "name": "outline-color",
          "value": "rgb(255, 0, 0)"
        }
      ]
    }
  ]
}
